+++ test/test_cluda/test_struct.py
@@ -5,16 +5,37 @@ import reikna.cluda as cluda
 import reikna.cluda.dtypes as dtypes
 
 
-def flatten_dtype(dtype, prefix=[]):
+def pytest_generate_tests(metafunc):
 
-    if dtype.names is None:
-        return [(prefix, dtype)]
-    else:
-        result = []
-        for name in dtype.names:
-            nested_dtype, offset = dtype.fields[name]
-            result += flatten_dtype(nested_dtype, prefix=prefix + [name])
-        return result
+    if 'dtype_to_align' in metafunc.funcargnames:
+
+        vals = []
+        ids = []
+
+        # numpy itemsize == 9, but on device it will be aligned to 4, so the total size will be 12
+        dtype = numpy.dtype([('val1', numpy.int32), ('val2', numpy.int32), ('pad', numpy.int8)])
+        vals.append(dtype)
+        ids.append("small_pad")
+
+        dtype_nested = numpy.dtype([
+            ('val1', numpy.int32), ('pad', numpy.int8)])
+        dtype = numpy.dtype([
+            ('val1', numpy.int32),
+            ('val2', numpy.int16),
+            ('nested', dtype_nested)])
+        vals.append(dtype)
+        ids.append("nested")
+
+        dtype_nested = numpy.dtype(dict(
+            names=['val1', 'pad'],
+            formats=[numpy.int8, numpy.int8]))
+        dtype = numpy.dtype(dict(
+            names=['pad', 'struct_arr', 'regular_arr'],
+            formats=[numpy.int32, numpy.dtype((dtype_nested, 2)), numpy.dtype((numpy.int16, 3))]))
+        vals.append(dtype)
+        ids.append("nested_array")
+
+        metafunc.parametrize('dtype_to_align', vals, ids=ids)
 
 
 def extract_field(arr, path):
@@ -24,10 +45,63 @@ def extract_field(arr, path):
         return extract_field(arr[path[0]], path[1:])
 
 
+def get_offsets_from_device(thr, dtype):
+    """
+    Returns a new data type object with the same fields as in ``dtype``
+    and the field offsets set by the compiler in ``thr``.
+    All existing offsets in ``dtype`` are ignored.
+    """
+
+    if dtype.names is None:
+        return dtype
+
+    struct = dtypes.ctype_module(dtype, ignore_alignment=True)
+
+    program = thr.compile(
+    """
+    #define my_offsetof(type, field) ((size_t)(&((type *)0)->field))
+
+    KERNEL void test(GLOBAL_MEM int *dest)
+    {
+      %for i, name in enumerate(dtype.names):
+      dest[${i}] = my_offsetof(${struct}, ${name});
+      %endfor
+      dest[${len(dtype.names)}] = sizeof(${struct});
+    }
+    """, render_kwds=dict(
+        struct=struct,
+        dtype=dtype))
+
+    offsets = thr.array(len(dtype.names) + 1, numpy.int32)
+    test = program.test
+    test(offsets, global_size=1)
+    offsets = offsets.get()
+
+    # Casting to Python ints, becase numpy ints as dtype offsets make it unhashable.
+    offsets = [int(offset) for offset in offsets]
+
+    aligned_dtypes = [
+        get_offsets_from_device(thr, dtype.fields[name][0])
+        for name in dtype.names]
+
+    return numpy.dtype(dict(
+        names=dtype.names,
+        formats=aligned_dtypes,
+        offsets=offsets[:-1],
+        itemsize=offsets[-1],
+        aligned=True))
+
+
+def test_align(thr, dtype_to_align):
+    aligned_dtype = dtypes.align(dtype_to_align)
+    empyric_dtype = get_offsets_from_device(thr, dtype_to_align)
+    assert aligned_dtype == empyric_dtype
+
+
 def check_struct_fill(thr, dtype):
     """
     Fill every field of the given ``dtype`` with its number and check the results.
-    This helps to detect issues with alignment in the struct.
+    This helps to detect issues with offsets in the struct.
     """
     struct = dtypes.ctype_module(dtype)
 
@@ -38,7 +112,7 @@ def check_struct_fill(thr, dtype):
       const SIZE_T i = get_global_id(0);
       ${struct} res;
 
-      %for i, field_info in enumerate(flatten_dtype(dtype)):
+      %for i, field_info in enumerate(dtypes.flatten_dtype(dtype)):
       res.${".".join(field_info[0])} = ${i};
       %endfor
 
@@ -46,8 +120,7 @@ def check_struct_fill(thr, dtype):
     }
     """, render_kwds=dict(
         struct=struct,
-        dtype=dtype,
-        flatten_dtype=flatten_dtype))
+        dtype=dtype))
 
     test = program.test
 
@@ -55,12 +128,12 @@ def check_struct_fill(thr, dtype):
     test(a_dev, global_size=128)
     a = a_dev.get()
 
-    for i, field_info in enumerate(flatten_dtype(dtype)):
+    for i, field_info in enumerate(dtypes.flatten_dtype(dtype)):
         path, _ = field_info
         assert (extract_field(a, path) == i).all()
 
 
-def test_hardcoded_alignment(thr):
+def test_hardcoded_offsets(thr):
     """
     Test the correctness of alignment for an explicit set of field offsets.
     """
@@ -69,18 +142,20 @@ def test_hardcoded_alignment(thr):
         names=['val1', 'pad'],
         formats=[numpy.int32, numpy.int8],
         offsets=[0, 4],
-        itemsize=8))
+        itemsize=8,
+        aligned=True))
 
     dtype = numpy.dtype(dict(
         names=['val1', 'val2', 'nested'],
         formats=[numpy.int32, numpy.int16, dtype_nested],
         offsets=[0, 4, 8],
-        itemsize=32))
+        itemsize=32,
+        aligned=True))
 
     check_struct_fill(thr, dtype)
 
 
-def test_adjusted_alignment(thr):
+def test_align(thr):
     """
     Test the correctness of alignment for field offsets adjusted automatically.
     """
@@ -93,7 +168,7 @@ def test_adjusted_alignment(thr):
         ('val2', numpy.int16),
         ('nested', dtype_nested)])
 
-    dtype = dtypes.adjust_alignment(thr, dtype)
+    dtype = dtypes.align(dtype)
 
     check_struct_fill(thr, dtype)
 
@@ -110,7 +185,7 @@ def test_nested_array(thr):
         names=['pad', 'struct_arr', 'regular_arr'],
         formats=[numpy.int32, numpy.dtype((dtype_nested, 2)), numpy.dtype((numpy.int16, 3))]))
 
-    dtype = dtypes.adjust_alignment(thr, dtype)
+    dtype = dtypes.align(dtype)
     struct = dtypes.ctype_module(dtype)
 
     program = thr.compile(
@@ -131,8 +206,7 @@ def test_nested_array(thr):
     }
     """, render_kwds=dict(
         struct=struct,
-        dtype=dtype,
-        flatten_dtype=flatten_dtype))
+        dtype=dtype))
 
     test = program.test
 
@@ -153,7 +227,7 @@ def test_structural_typing(some_thr):
     Checks that ``ctype_module`` for equal dtype objects result in the same module object
     (which means that these two types will actually be rendered as a single type).
     """
-    dtype = numpy.dtype([('val1', numpy.int32), ('val2', numpy.float32)])
+    dtype = dtypes.align(numpy.dtype([('val1', numpy.int32), ('val2', numpy.float32)]))
 
     struct1 = dtypes.ctype_module(dtype)
     struct2 = dtypes.ctype_module(dtype)
@@ -191,10 +265,11 @@ def test_structural_typing_nested(some_thr):
     In other words, a nested dtype gets represented by the same module as
     an equal top-level dtype.
     """
-    dtype_nested = numpy.dtype([('val1', numpy.int32), ('val2', numpy.float32)])
-    dtype = numpy.dtype([
+    dtype_nested = dtypes.align(
+        numpy.dtype([('val1', numpy.int32), ('val2', numpy.float32)]))
+    dtype = dtypes.align(numpy.dtype([
         ('val1', numpy.int32), ('val2', numpy.float32),
-        ('nested', dtype_nested)])
+        ('nested', dtype_nested)]))
 
     struct_nested = dtypes.ctype_module(dtype_nested)
     struct = dtypes.ctype_module(dtype)